+++ tests/variational/ucj_test.py
@@ -72,7 +72,7 @@ def test_t_amplitudes_roundtrip():
 
     rng = np.random.default_rng()
 
-    t2 = ffsim.random.random_t2_amplitudes(norb, nocc)
+    t2 = ffsim.random.random_t2_amplitudes(norb, nocc, dtype=float)
     t1 = rng.standard_normal((nocc, norb - nocc))
 
     operator = ffsim.UCJOperator.from_t_amplitudes(t2, t1_amplitudes=t1)
@@ -176,7 +176,7 @@ def test_real_ucj_t_amplitudes_roundtrip():
 
     rng = np.random.default_rng()
 
-    t2 = ffsim.random.random_t2_amplitudes(norb, nocc)
+    t2 = ffsim.random.random_t2_amplitudes(norb, nocc, dtype=float)
     t1 = rng.standard_normal((nocc, norb - nocc))
 
     operator = ffsim.RealUCJOperator.from_t_amplitudes(t2, t1_amplitudes=t1)
